reorganize chessboard sub-component code into smaller meaningful files/functions
